python-backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pyresparser==1.0.6
spacy==3.7.2
nltk==3.8.1
python-docx==0.8.11
PyPDF2==3.0.1
pdfplumber==0.10.3
pymupdf==1.23.26
werkzeug==2.3.7
